[feat] 내가 쓴 후기 리스트 조회 API 구현 및 자체 qa를 통한 전체적인 보수, 리펙토링 작업

+++ src/main/java/org/sopt/pawkey/backendapi/domain/post/application/service/PostLikeService.java
@@ -1,8 +1,8 @@
 package org.sopt.pawkey.backendapi.domain.post.application.service;
 
+import org.sopt.pawkey.backendapi.domain.post.api.dto.response.PostLikeResponse;
 import org.sopt.pawkey.backendapi.domain.post.domain.repository.PostLikeRepository;
-import org.sopt.pawkey.backendapi.domain.post.exception.PostLikeBusinessException;
-import org.sopt.pawkey.backendapi.domain.post.exception.PostLikeErrorCode;
+import org.sopt.pawkey.backendapi.domain.post.domain.repository.PostRepository;
 import org.sopt.pawkey.backendapi.domain.post.infra.persistence.entity.PostEntity;
 import org.sopt.pawkey.backendapi.domain.post.infra.persistence.entity.PostLikeEntity;
 import org.sopt.pawkey.backendapi.domain.user.infra.persistence.entity.UserEntity;
@@ -15,33 +15,24 @@ import lombok.RequiredArgsConstructor;
 public class PostLikeService {
 
 	private final PostLikeRepository postLikeRepository;
+	private final PostRepository postRepository;
 
-	public void like(final UserEntity user, final PostEntity post) {
-		// validateNotSelfLike(user, post);
+	public PostLikeResponse toggleLike(final UserEntity user, final PostEntity post) {
+		boolean isLiked = postLikeRepository.existsByUserIdAndPostId(user.getUserId(), post.getPostId());
 
-		if (postLikeRepository.existsByUserIdAndPostId(user.getUserId(), post.getPostId())) {
-			throw new PostLikeBusinessException(PostLikeErrorCode.DUPLICATE_LIKE);
-		}
+		if (isLiked) {
+			postLikeRepository.deleteByUserIdAndPostId(user.getUserId(), post.getPostId());
+			postRepository.decreaseLikeCount(post.getPostId());
 
-		postLikeRepository.save(PostLikeEntity.builder()
-			.post(post)
-			.user(user)
-			.build());
-	}
+			return PostLikeResponse.of(PostLikeResponse.CANCEL_SUCCESS);
+		} else {
+			postLikeRepository.save(PostLikeEntity.builder()
+				.post(post)
+				.user(user)
+				.build());
+			postRepository.increaseLikeCount(post.getPostId());
 
-	public void cancelLike(UserEntity user, PostEntity post) {
-		// validateNotSelfLike(user, post);
-
-		long deletedCount = postLikeRepository.deleteByUserIdAndPostId(user.getUserId(), post.getPostId());
-
-		if (deletedCount == 0) {
-			throw new PostLikeBusinessException(PostLikeErrorCode.LIKE_NOT_FOUND);
+			return PostLikeResponse.of(PostLikeResponse.LIKE_SUCCESS);
 		}
 	}
-
-	// private void validateNotSelfLike(UserEntity user, PostEntity post) {
-	// 	if (post.getUser().getUserId().equals(user.getUserId())) {
-	// 		throw new PostLikeBusinessException(PostLikeErrorCode.CANNOT_LIKE_OWN_POST);
-	// 	}
-	// }
 }
